+++ fiktion-core/src/commonMain/kotlin/dev/s7a/fiktion/PrefixedGenerationSpec.kt
@@ -12,6 +12,7 @@ internal fun DefaultGenerationSpec<*>.prefixedBy(prefix: List<PathRuleSegment>):
         nullProbability = nullProbability,
         defaultProbability = defaultProbability,
         automaticallyGenerates = automaticallyGenerates,
+        defaultGenerates = defaultGenerates,
         precedence = precedence,
     )